6.20ms -> 6.15ms. Enabling secret -Wsign-conversion flag to find all implicit unsigned int to int conversions

+++ src/board/sdl.cpp
@@ -19,7 +19,7 @@ void renderBoardSdl(const Board board, SDL_Renderer *renderer, SDL_Texture *text
   const auto &[input, width, height] = board;
 
   static_assert(sizeof(input[0]) == sizeof(Uint32));
-  SDL_UpdateTexture(texture, NULL, &input[0], width * sizeof(Uint32));
+  SDL_UpdateTexture(texture, NULL, &input[0], (long unsigned int)width * sizeof(Uint32));
 
   SDL_RenderCopy(renderer, texture, nullptr, nullptr);
   SDL_RenderPresent(renderer);
